Slimmer #includes behaviour
The goal of the commit is to make #includes slimmer -
only join the table required to satisfy the referenced tables and
preload all other associations independently.

Context

The Active Record #includes method results in different behaviour
depending on the other parts of the query:
it may execute multiple queries (one per included association) or
build a single JOIN query to load everything at once.

Before the commit, it was "all or nothing" behind an eager load flag -
one referenced table was enough to transform the whole query into one big JOIN.

However, it's not always necessary to join all the tables
except the ones with conditions.
Normally, we could avoid joining everything (which could be performance-heavy),
and execute a few queries instead.

Idea

The main idea is to remove referenced tables
(which are #references, #where, #order or copied by #and, #or, #merge)
from the final join query.
So we would LEFT OUTER JOINed only includes tables with conditions on the query.
At the same time,
we have to bring non-referenced tables back to separate query preloads.

Backward compatibility

1) Includes and table names don't always match so we use reflections.

We try to match all possible includes table names against references
including inner and left outer joins.
If there's an unreliable match (like no reflection for association),
we're inclined to JOIN the associations in order to not break user applications.

All current tests are green without changes inside them.

2) The commit certainly breaks user code with the plain string where or select:
```
Author.includes(:company, :books).where(company: { title: "The Office" })
                                 .where("books.title = ?", "Design Patterns")
                                 .select("books.title")
```
We do not want to introduce any flags or scopes to support this behaviour
because in docs there're mentions to prohibit string causes without references:
https://api.rubyonrails.org/v7.0.4.2/classes/ActiveRecord/QueryMethods.html#method-i-includes
https://api.rubyonrails.org/v7.0.4.2/classes/ActiveRecord/QueryMethods.html#method-i-references

People who don't know it have to start using it.

Future improvements

1) Includes are not always plain - they may contain several levels inside a hash.

In the commit, we handle several includes
as separate trees starting from the top-level root.
If there's any reference to possible includes table,
we'll mark this whole tree as needed to JOINed.

It's a safe bet which could be improved in future
with a more fine-grained reference marking algorithm.

2) The commit's tree traversal is average performant at best.
Several intermediate structures and tree walks could be removed -
we prefer a simple algorithm at the moment to show the idea.